renamed select_down_sample to select_by_index
