+++ src/Open3D/Geometry/TriangleMesh.cpp
@@ -1537,6 +1537,112 @@ void TriangleMesh::RemoveVerticesByMask(const std::vector<bool> &vertex_mask) {
     RemoveTrianglesByMask(triangle_mask);
 }
 
+std::shared_ptr<TriangleMesh> TriangleMesh::SelectByIndex(
+        const std::vector<size_t> &indices) const {
+    if (HasTriangleUvs()) {
+        utility::LogWarning(
+                "[SelectByIndices] This mesh contains triangle uvs that are "
+                "not handled in this function");
+    }
+    auto output = std::make_shared<TriangleMesh>();
+    bool has_triangle_normals = HasTriangleNormals();
+    bool has_vertex_normals = HasVertexNormals();
+    bool has_vertex_colors = HasVertexColors();
+    // For each vertex, list face indices.
+    std::vector<std::vector<int>> vertex_to_triangle_temp(vertices_.size());
+    int triangle_id = 0;
+    for (auto trangle : triangles_) {
+        for (int i = 0; i < 3; i++)
+            vertex_to_triangle_temp[trangle(i)].push_back(triangle_id);
+        triangle_id++;
+    }
+    // Remove face indices of vertex_to_triangle_temp
+    // if it does not correspond to selected vertices
+    std::vector<std::vector<int>> vertex_to_triangle(vertices_.size());
+    for (auto vertex_id : indices) {
+        vertex_to_triangle[vertex_id] = vertex_to_triangle_temp[vertex_id];
+    }
+    // Make a triangle_to_vertex using vertex_to_triangle
+    std::vector<std::vector<int>> triangle_to_vertex(triangles_.size());
+    int vertex_id = 0;
+    for (auto face_ids : vertex_to_triangle) {
+        for (auto face_id : face_ids)
+            triangle_to_vertex[face_id].push_back(vertex_id);
+        vertex_id++;
+    }
+    // Only a face with three selected points contributes to mark
+    // mask_observed_vertex.
+    std::vector<bool> mask_observed_vertex(vertices_.size());
+    for (auto vertex_ids : triangle_to_vertex) {
+        if ((int)vertex_ids.size() == 3)
+            for (int i = 0; i < 3; i++)
+                mask_observed_vertex[vertex_ids[i]] = true;
+    }
+    // Rename vertex id based on selected points
+    std::vector<int> new_vertex_id(vertices_.size());
+    for (size_t i = 0, cnt = 0; i < mask_observed_vertex.size(); i++) {
+        if (mask_observed_vertex[i]) {
+            new_vertex_id[i] = int(cnt);
+            cnt++;
+        }
+    }
+    // Push a triangle that has 3 selected vertices.
+    triangle_id = 0;
+    for (auto vertex_ids : triangle_to_vertex) {
+        if ((int)vertex_ids.size() == 3) {
+            Eigen::Vector3i new_face;
+            for (int i = 0; i < 3; i++)
+                new_face(i) = new_vertex_id[triangles_[triangle_id][i]];
+            output->triangles_.push_back(new_face);
+            if (has_triangle_normals)
+                output->triangle_normals_.push_back(
+                        triangle_normals_[triangle_id]);
+        }
+        triangle_id++;
+    }
+    // Push marked vertex.
+    for (size_t i = 0; i < mask_observed_vertex.size(); i++) {
+        if (mask_observed_vertex[i]) {
+            output->vertices_.push_back(vertices_[i]);
+            if (has_vertex_normals)
+                output->vertex_normals_.push_back(vertex_normals_[i]);
+            if (has_vertex_colors)
+                output->vertex_colors_.push_back(vertex_colors_[i]);
+        }
+    }
+    output->RemoveDuplicatedVertices();
+    output->RemoveDuplicatedTriangles();
+    output->RemoveUnreferencedVertices();
+    output->RemoveDegenerateTriangles();
+    utility::LogDebug(
+            "Triangle mesh sampled from {:d} vertices and {:d} triangles to "
+            "{:d} vertices and {:d} triangles.",
+            (int)vertices_.size(), (int)triangles_.size(),
+            (int)output->vertices_.size(), (int)output->triangles_.size());
+    return output;
+}
+
+std::shared_ptr<TriangleMesh> TriangleMesh::Crop(
+        const AxisAlignedBoundingBox &bbox) const {
+    if (bbox.IsEmpty()) {
+        utility::LogError(
+                "[CropTriangleMesh] AxisAlignedBoundingBox either has zeros "
+                "size, or has wrong bounds.");
+    }
+    return SelectByIndex(bbox.GetPointIndicesWithinBoundingBox(vertices_));
+}
+
+std::shared_ptr<TriangleMesh> TriangleMesh::Crop(
+        const OrientedBoundingBox &bbox) const {
+    if (bbox.IsEmpty()) {
+        utility::LogError(
+                "[CropTriangleMesh] AxisAlignedBoundingBox either has zeros "
+                "size, or has wrong bounds.");
+        return std::make_shared<TriangleMesh>();
+    }
+    return SelectByIndex(bbox.GetPointIndicesWithinBoundingBox(vertices_));
+}
+
 std::unordered_map<Eigen::Vector2i,
                    double,
                    utility::hash_eigen::hash<Eigen::Vector2i>>
