+++ src/Open3D/IO/ClassIO/PointCloudIO.cpp
@@ -110,7 +110,7 @@ bool ReadPointCloud(const std::string &filename,
                       (int)pointcloud.points_.size());
     if (remove_nan_points || remove_infinite_points) {
         pointcloud.RemoveNonFinitePoints(remove_nan_points,
-                                          remove_infinite_points);
+                                         remove_infinite_points);
     }
     return success;
 }
