+++ src/Open3D/Visualization/Utility/SelectionPolygon.cpp
@@ -207,24 +207,24 @@ SelectionPolygon::CreateSelectionPolygonVolume(const ViewControl &view) {
 std::shared_ptr<geometry::PointCloud>
 SelectionPolygon::CropPointCloudInRectangle(const geometry::PointCloud &input,
                                             const ViewControl &view) {
-    return input.SelectDownSample(CropInRectangle(input.points_, view));
+    return input.SelectByIndex(CropInRectangle(input.points_, view));
 }
 
 std::shared_ptr<geometry::PointCloud> SelectionPolygon::CropPointCloudInPolygon(
         const geometry::PointCloud &input, const ViewControl &view) {
-    return input.SelectDownSample(CropInPolygon(input.points_, view));
+    return input.SelectByIndex(CropInPolygon(input.points_, view));
 }
 
 std::shared_ptr<geometry::TriangleMesh>
 SelectionPolygon::CropTriangleMeshInRectangle(
         const geometry::TriangleMesh &input, const ViewControl &view) {
-    return input.SelectDownSample(CropInRectangle(input.vertices_, view));
+    return input.SelectByIndex(CropInRectangle(input.vertices_, view));
 }
 
 std::shared_ptr<geometry::TriangleMesh>
 SelectionPolygon::CropTriangleMeshInPolygon(const geometry::TriangleMesh &input,
                                             const ViewControl &view) {
-    return input.SelectDownSample(CropInPolygon(input.vertices_, view));
+    return input.SelectByIndex(CropInPolygon(input.vertices_, view));
 }
 
 std::vector<size_t> SelectionPolygon::CropInRectangle(
