+++ src/Python/open3d_pybind/geometry/pointcloud.cpp
@@ -67,7 +67,7 @@ void pybind_pointcloud(py::module &m) {
             .def("paint_uniform_color",
                  &geometry::PointCloud::PaintUniformColor, "color"_a,
                  "Assigns each point in the PointCloud the same color.")
-            .def("select_down_sample", &geometry::PointCloud::SelectDownSample,
+            .def("select_by_index", &geometry::PointCloud::SelectByIndex,
                  "Function to select points from input pointcloud into output "
                  "pointcloud.",
                  "indices"_a, "invert"_a = false)
@@ -223,7 +223,7 @@ void pybind_pointcloud(py::module &m) {
             m, "PointCloud", "paint_uniform_color",
             {{"color", "RGB color for the PointCloud."}});
     docstring::ClassMethodDocInject(
-            m, "PointCloud", "select_down_sample",
+            m, "PointCloud", "select_by_index",
             {{"indices", "Indices of points to be selected."},
              {"invert",
               "Set to ``True`` to invert the selection of indices."}});
